server/requirements.txt
flask
fastapi
uvicorn
python-multipart
pydantic-settings
bcrypt
python-dotenv
python-jose[cryptography]
pymysql